refactor(pgstac-rs): reorganize modules; drop dev artifacts and dead role
Groups the flat crate root into db/ read/ load/ json/ api/ python/ and finishes the
rustac-native + cleanup pass. Precheck stays a library method, not yet wired into load.

pgstac-rs:
- reorg: flat root modules into db/ read/ load/ json/ api/ python/ (public API unchanged)
- search CLI dedup: route the itemcollection command through stac::api ItemsClient::search;
  delete SearchSource and page_to_feature_collection; map bad search tokens to Error::InvalidToken
- remove the profiling feature, memprofile example, and dhat dep (preserved on the
  archive/memprofile branch); drop the lib.rs dhat shim
- drop the EXPERIMENTAL hedge from the precheck docs; comments describe current behavior only

pgstac (SQL):
- remove the unused pgstac_load up-privilege role (added this cycle, never released); the
  SECURITY DEFINER write functions remain the only write path through the privilege wall
- add a pgtap proof-test that the wall holds: pgstac_ingest cannot directly INSERT/UPDATE/DELETE
  items, partition_stats, or item_fragments
- regenerate pgstac.sql
