+++ src/pgstac-rs/src/api/mod.rs
@@ -1,0 +1,4 @@
+//! The rustac-native `stac::api` client-trait implementations over [`PgstacPool`](crate::PgstacPool).
+
+#[cfg(feature = "pool")]
+pub(crate) mod stac_api;
