+++ src/pgstac-rs/src/json/canonical.rs
