+++ src/pgstac-rs/src/json/mod.rs
@@ -1,0 +1,6 @@
+//! Low-level JSON and column-conversion helpers (canonical hashing, raw jsonb, geometry, datetime).
+
+pub mod canonical;
+pub mod geom;
+pub mod rawjson;
+pub mod temporal;
