+++ src/pgstac-rs/src/load/field_registry.rs
